refactor(cozy-mespapiers-lib): Move all components PlaceholderList
Move all components PlaceholderList on a new folder
to improve understanding of the lib
