+++ packages/cozy-mespapiers-lib/src/components/Placeholders/PlaceholderListModal/PlaceholderThemesList.jsx
