+++ packages/cozy-mespapiers-lib/src/components/Placeholders/PlaceholderListModal/PlaceholdersList.jsx
